[0.4.0] staking pool: Add helper methods for delegators
