+++ staking-pool/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "staking-pool"
-version = "0.3.0"
+version = "0.4.0"
 authors = ["Near Inc <hello@near.org>"]
 edition = "2018"
 publish = false
