+++ staking-pool/src/lib.rs
@@ -10,6 +10,8 @@ use near_sdk::{
 };
 use uint::construct_uint;
 
+mod internal;
+
 /// The amount of gas given to complete `vote` call.
 const VOTE_GAS: u64 = 100_000_000_000_000;
 
@@ -199,14 +201,14 @@ impl StakingContract {
             paused: false,
         };
         // Staking with the current pool to make sure the staking key is valid.
-        this.restake();
+        this.internal_restake();
         this
     }
 
     /// Distributes rewards and restakes if needed.
     pub fn ping(&mut self) {
         if self.internal_ping() {
-            self.restake();
+            self.internal_restake();
         }
     }
 
@@ -215,61 +217,61 @@ impl StakingContract {
     pub fn deposit(&mut self) {
         let need_to_restake = self.internal_ping();
 
-        let account_id = env::predecessor_account_id();
-        let mut account = self.get_inner_account(&account_id);
-        let amount = env::attached_deposit();
-        account.unstaked += amount;
-        self.save_inner_account(&account_id, &account);
-        self.last_total_balance += amount;
-
-        env::log(
-            format!(
-                "@{} deposited {}. New unstaked balance is {}",
-                account_id, amount, account.unstaked
-            )
-            .as_bytes(),
-        );
+        let _amount = self.internal_deposit();
 
         if need_to_restake {
-            self.restake();
+            self.internal_restake();
+        }
+    }
+
+    /// Deposits the attached amount into the inner account of the predecessor and stakes it.
+    #[payable]
+    pub fn deposit_and_stake(&mut self) {
+        let _need_to_restake = self.internal_ping();
+
+        let amount = self.internal_deposit();
+        self.internal_stake(amount);
+
+        self.internal_restake();
+    }
+
+    /// Withdraws the entire unstaked balance from the predecessor account.
+    /// It's only allowed if the `unstake` action was not performed in the four most recent epochs.
+    pub fn withdraw_all(&mut self) {
+        let need_to_restake = self.internal_ping();
+
+        let account_id = env::predecessor_account_id();
+        let account = self.internal_get_account(&account_id);
+        self.internal_withdraw(account.unstaked);
+
+        if need_to_restake {
+            self.internal_restake();
         }
     }
 
     /// Withdraws the non staked balance for given account.
-    /// It's only allowed if the `unstake` action was not performed in the three most recent epochs.
+    /// It's only allowed if the `unstake` action was not performed in the four most recent epochs.
     pub fn withdraw(&mut self, amount: U128) {
         let need_to_restake = self.internal_ping();
 
-        let amount = amount.into();
-        assert!(amount > 0, "Withdrawal amount should be positive");
-
-        let account_id = env::predecessor_account_id();
-        let mut account = self.get_inner_account(&account_id);
-        assert!(
-            account.unstaked >= amount,
-            "Not enough unstaked balance to withdraw"
-        );
-        assert!(
-            account.unstaked_available_epoch_height <= env::epoch_height(),
-            "The unstaked balance is not yet available due to unstaking delay"
-        );
-        account.unstaked -= amount;
-        self.save_inner_account(&account_id, &account);
-
-        env::log(
-            format!(
-                "@{} withdrawing {}. New unstaked balance is {}",
-                account_id, amount, account.unstaked
-            )
-            .as_bytes(),
-        );
-
-        Promise::new(account_id).transfer(amount);
-        self.last_total_balance -= amount;
+        let amount: Balance = amount.into();
+        self.internal_withdraw(amount);
 
         if need_to_restake {
-            self.restake();
+            self.internal_restake();
         }
+    }
+
+    /// Stakes all available unstaked balance from the inner account of the predecessor.
+    pub fn stake_all(&mut self) {
+        // Stake action always restakes
+        let _need_to_restake = self.internal_ping();
+
+        let account_id = env::predecessor_account_id();
+        let account = self.internal_get_account(&account_id);
+        self.internal_stake(account.unstaked);
+
+        self.internal_restake();
     }
 
     /// Stakes the given amount from the inner account of the predecessor.
@@ -279,143 +281,36 @@ impl StakingContract {
         let _need_to_restake = self.internal_ping();
 
         let amount: Balance = amount.into();
-        assert!(amount > 0, "Staking amount should be positive");
+        self.internal_stake(amount);
+
+        self.internal_restake();
+    }
+
+    /// Unstakes all staked balance from the inner account of the predecessor.
+    /// The new total unstaked balance will be available for withdrawal in four epochs.
+    pub fn unstake_all(&mut self) {
+        // Unstake action always restakes
+        let _need_to_restake = self.internal_ping();
 
         let account_id = env::predecessor_account_id();
-        let mut account = self.get_inner_account(&account_id);
+        let account = self.internal_get_account(&account_id);
+        let amount = self.staked_amount_from_num_shares_rounded_down(account.stake_shares);
+        self.inner_unstake(amount);
 
-        // Calculate the number of "stake" shares that the account will receive for staking the
-        // given amount.
-        let num_shares = self.num_shares_from_staked_amount_rounded_down(amount);
-        assert!(
-            num_shares > 0,
-            "The calculated number of \"stake\" shares received for staking should be positive"
-        );
-        // The amount of tokens the account will be charged from the unstaked balance.
-        // Rounded down to avoid overcharging the account to guarantee that the account can always
-        // unstake at least the same amount as staked.
-        let charge_amount = self.staked_amount_from_num_shares_rounded_down(num_shares);
-        assert!(
-            charge_amount > 0,
-            "Invariant violation. Calculated staked amount must be positive, because \"stake\" share price should be at least 1"
-        );
-
-        assert!(
-            account.unstaked >= charge_amount,
-            "Not enough unstaked balance to stake"
-        );
-        account.unstaked -= charge_amount;
-        account.stake_shares += num_shares;
-        self.save_inner_account(&account_id, &account);
-
-        // The staked amount that will be added to the total to guarantee the "stake" share price
-        // never decreases. The difference between `stake_amount` and `charge_amount` is paid
-        // from the allocated STAKE_SHARE_PRICE_GUARANTEE_FUND.
-        let stake_amount = self.staked_amount_from_num_shares_rounded_up(num_shares);
-
-        self.total_staked_balance += stake_amount;
-        self.total_stake_shares += num_shares;
-
-        env::log(
-            format!(
-                "@{} staking {}. Received {} new staking shares. Total {} unstaked balance and {} staking shares",
-                account_id, charge_amount, num_shares, account.unstaked, account.stake_shares
-            )
-            .as_bytes(),
-        );
-        env::log(
-            format!(
-                "Contract total staked balance is {}. Total number of shares {}",
-                self.total_staked_balance, self.total_stake_shares
-            )
-            .as_bytes(),
-        );
-
-        self.restake();
+        self.internal_restake();
     }
 
     /// Unstakes the given amount from the inner account of the predecessor.
     /// The inner account should have enough staked balance.
-    /// The new total unstaked balance will be available for withdrawal in 3 epochs.
+    /// The new total unstaked balance will be available for withdrawal in four epochs.
     pub fn unstake(&mut self, amount: U128) {
         // Unstake action always restakes
         let _need_to_restake = self.internal_ping();
 
         let amount: Balance = amount.into();
-        assert!(amount > 0, "Unstaking amount should be positive");
+        self.inner_unstake(amount);
 
-        let account_id = env::predecessor_account_id();
-        let mut account = self.get_inner_account(&account_id);
-
-        assert!(
-            self.total_staked_balance > 0,
-            "The contract doesn't have staked balance"
-        );
-        // Calculate the number of shares required to unstake the given amount.
-        // NOTE: The number of shares the account will pay is rounded up.
-        let num_shares = self.num_shares_from_staked_amount_rounded_up(amount);
-        assert!(
-            num_shares > 0,
-            "Invariant violation. The calculated number of \"stake\" shares for unstaking should be positive"
-        );
-        assert!(
-            account.stake_shares >= num_shares,
-            "Not enough staked balance to unstake"
-        );
-
-        // Calculating the amount of tokens the account will receive by unstaking the corresponding
-        // number of "stake" shares, rounding up.
-        let receive_amount = self.staked_amount_from_num_shares_rounded_up(num_shares);
-        assert!(
-            receive_amount > 0,
-            "Invariant violation. Calculated staked amount must be positive, because \"stake\" share price should be at least 1"
-        );
-
-        account.stake_shares -= num_shares;
-        account.unstaked += receive_amount;
-        account.unstaked_available_epoch_height = env::epoch_height() + NUM_EPOCHS_TO_UNLOCK;
-        self.save_inner_account(&account_id, &account);
-
-        // The amount tokens that will be unstaked from the total to guarantee the "stake" share
-        // price never decreases. The difference between `receive_amount` and `unstake_amount` is
-        // paid from the allocated STAKE_SHARE_PRICE_GUARANTEE_FUND.
-        let unstake_amount = self.staked_amount_from_num_shares_rounded_down(num_shares);
-
-        self.total_staked_balance -= unstake_amount;
-        self.total_stake_shares -= num_shares;
-
-        env::log(
-            format!(
-                "@{} unstaking {}. Spent {} staking shares. Total {} unstaked balance and {} staking shares",
-                account_id, receive_amount, num_shares, account.unstaked, account.stake_shares
-            )
-                .as_bytes(),
-        );
-        env::log(
-            format!(
-                "Contract total staked balance is {}. Total number of shares {}",
-                self.total_staked_balance, self.total_stake_shares
-            )
-            .as_bytes(),
-        );
-
-        self.restake();
-    }
-
-    /// Restakes the current `total_staked_balance` again.
-    fn restake(&mut self) {
-        if self.paused {
-            return;
-        }
-        // Stakes with the staking public key. If the public key is invalid the entire function
-        // call will be rolled back.
-        Promise::new(env::current_account_id())
-            .stake(self.total_staked_balance, self.stake_public_key.clone())
-            .then(ext_self::on_stake_action(
-                &env::current_account_id(),
-                NO_DEPOSIT,
-                ON_STAKE_ACTION_GAS,
-            ));
+        self.internal_restake();
     }
 
     /****************/
@@ -472,7 +367,7 @@ impl StakingContract {
 
     /// Returns human readable representation of the account for the given account ID.
     pub fn get_account(&self, account_id: AccountId) -> HumanReadableAccount {
-        let account = self.get_inner_account(&account_id);
+        let account = self.internal_get_account(&account_id);
         HumanReadableAccount {
             account_id,
             unstaked_balance: account.unstaked.into(),
@@ -536,7 +431,7 @@ impl StakingContract {
         // When updating the staking key, the contract has to restake.
         let _need_to_restake = self.internal_ping();
         self.stake_public_key = stake_public_key.into();
-        self.restake();
+        self.internal_restake();
     }
 
     /// Owner's method.
@@ -548,7 +443,7 @@ impl StakingContract {
         let need_to_restake = self.internal_ping();
         self.reward_fee_fraction = reward_fee_fraction;
         if need_to_restake {
-            self.restake();
+            self.internal_restake();
         }
     }
 
@@ -583,154 +478,7 @@ impl StakingContract {
 
         self.internal_ping();
         self.paused = false;
-        self.restake();
-    }
-
-    /********************/
-    /* Internal methods */
-    /********************/
-
-    /// Asserts that the method was called by the owner.
-    fn assert_owner(&self) {
-        assert_eq!(
-            env::predecessor_account_id(),
-            self.owner_id,
-            "Can only be called by the owner"
-        );
-    }
-
-    /// Distributes rewards after the new epoch. It's automatically called before every action.
-    /// Returns true if the current epoch height is different from the last epoch height.
-    fn internal_ping(&mut self) -> bool {
-        let epoch_height = env::epoch_height();
-        if self.last_epoch_height == epoch_height {
-            return false;
-        }
-        self.last_epoch_height = epoch_height;
-
-        // New total amount (both locked and unlocked balances).
-        // NOTE: We need to subtract `attached_deposit` in case `ping` called from `deposit` call
-        // since the attached deposit gets included in the `account_balance`, and we have not
-        // accounted it yet.
-        let total_balance =
-            env::account_locked_balance() + env::account_balance() - env::attached_deposit();
-
-        assert!(
-            total_balance >= self.last_total_balance,
-            "The new total balance should not be less than the old total balance"
-        );
-        let total_reward = total_balance - self.last_total_balance;
-        if total_reward > 0 {
-            // The validation fee that the contract owner takes.
-            let owners_fee = self.reward_fee_fraction.multiply(total_reward);
-
-            // Distributing the remaining reward to the delegators first.
-            let remaining_reward = total_reward - owners_fee;
-            self.total_staked_balance += remaining_reward;
-
-            // Now buying "stake" shares for the contract owner at the new share price.
-            let num_shares = self.num_shares_from_staked_amount_rounded_down(owners_fee);
-            if num_shares > 0 {
-                // Updating owner's inner account
-                let owner_id = self.owner_id.clone();
-                let mut account = self.get_inner_account(&owner_id);
-                account.stake_shares += num_shares;
-                self.save_inner_account(&owner_id, &account);
-                // Increasing the total amount of "stake" shares.
-                self.total_stake_shares += num_shares;
-            }
-            // Increasing the total staked balance by the owners fee, no matter whether the owner
-            // received any shares or not.
-            self.total_staked_balance += owners_fee;
-
-            env::log(
-                format!(
-                    "Epoch {}: Contract received total rewards of {} tokens. New total staked balance is {}. Total number of shares {}",
-                    epoch_height, total_reward, self.total_staked_balance, self.total_stake_shares,
-                )
-                    .as_bytes(),
-            );
-            if num_shares > 0 {
-                env::log(format!("Total rewards fee is {} stake shares.", num_shares).as_bytes());
-            }
-        }
-
-        self.last_total_balance = total_balance;
-        true
-    }
-
-    /// Returns the number of "stake" shares rounded down corresponding to the given staked balance
-    /// amount.
-    ///
-    /// price = total_staked / total_shares
-    /// Price is fixed
-    /// (total_staked + amount) / (total_shares + num_shares) = total_staked / total_shares
-    /// (total_staked + amount) * total_shares = total_staked * (total_shares + num_shares)
-    /// amount * total_shares = total_staked * num_shares
-    /// num_shares = amount * total_shares / total_staked
-    fn num_shares_from_staked_amount_rounded_down(&self, amount: Balance) -> NumStakeShares {
-        assert!(
-            self.total_staked_balance > 0,
-            "The total staked balance can't be 0"
-        );
-        (U256::from(self.total_stake_shares) * U256::from(amount)
-            / U256::from(self.total_staked_balance))
-        .as_u128()
-    }
-
-    /// Returns the number of "stake" shares rounded up corresponding to the given staked balance
-    /// amount.
-    ///
-    /// Rounding up division of `a / b` is done using `(a + b - 1) / b`.
-    fn num_shares_from_staked_amount_rounded_up(&self, amount: Balance) -> NumStakeShares {
-        assert!(
-            self.total_staked_balance > 0,
-            "The total staked balance can't be 0"
-        );
-        ((U256::from(self.total_stake_shares) * U256::from(amount)
-            + U256::from(self.total_staked_balance - 1))
-            / U256::from(self.total_staked_balance))
-        .as_u128()
-    }
-
-    /// Returns the staked amount rounded down corresponding to the given number of "stake" shares.
-    fn staked_amount_from_num_shares_rounded_down(&self, num_shares: NumStakeShares) -> Balance {
-        assert!(
-            self.total_stake_shares > 0,
-            "The total number of stake shares can't be 0"
-        );
-        (U256::from(self.total_staked_balance) * U256::from(num_shares)
-            / U256::from(self.total_stake_shares))
-        .as_u128()
-    }
-
-    /// Returns the staked amount rounded up corresponding to the given number of "stake" shares.
-    ///
-    /// Rounding up division of `a / b` is done using `(a + b - 1) / b`.
-    fn staked_amount_from_num_shares_rounded_up(&self, num_shares: NumStakeShares) -> Balance {
-        assert!(
-            self.total_stake_shares > 0,
-            "The total number of stake shares can't be 0"
-        );
-        ((U256::from(self.total_staked_balance) * U256::from(num_shares)
-            + U256::from(self.total_stake_shares - 1))
-            / U256::from(self.total_stake_shares))
-        .as_u128()
-    }
-
-    /// Inner method to get the given account or a new default value account.
-    fn get_inner_account(&self, account_id: &AccountId) -> Account {
-        self.accounts.get(account_id).unwrap_or_default()
-    }
-
-    /// Inner method to save the given account for a given account ID.
-    /// If the account balances are 0, the account is deleted instead to release storage.
-    fn save_inner_account(&mut self, account_id: &AccountId, account: &Account) {
-        if account.unstaked > 0 || account.stake_shares > 0 {
-            self.accounts.insert(account_id, &account);
-        } else {
-            self.accounts.remove(account_id);
-        }
+        self.internal_restake();
     }
 }
 
@@ -843,7 +591,7 @@ mod tests {
             zero_fee(),
         );
         emulator.update_context(bob(), 0);
-        emulator.contract.restake();
+        emulator.contract.internal_restake();
         let receipts = env::created_receipts();
         assert_eq!(receipts.len(), 2);
         // Mocked Receipt fields are private, so can't check directly.
@@ -1007,6 +755,44 @@ mod tests {
         assert!(emulator
             .contract
             .is_account_unstaked_balance_available(bob()),);
+    }
+
+    #[test]
+    fn test_stake_all_unstake_all() {
+        let mut emulator = Emulator::new(
+            owner(),
+            "KuTCtARNzxZQ3YvXDeLjx83FDqxv2SdQTSbiq876zR7".to_string(),
+            zero_fee(),
+        );
+        let deposit_amount = ntoy(1_000_000);
+        emulator.update_context(bob(), deposit_amount);
+        emulator.contract.deposit_and_stake();
+        emulator.amount += deposit_amount;
+        emulator.simulate_stake_call();
+        assert_eq!(
+            emulator.contract.get_account_staked_balance(bob()).0,
+            deposit_amount
+        );
+        assert_eq_in_near!(emulator.contract.get_account_unstaked_balance(bob()).0, 0);
+        let locked_amount = emulator.locked_amount;
+
+        // 10 epochs later, unstake all.
+        emulator.skip_epochs(10);
+        // Overriding rewards
+        emulator.locked_amount = locked_amount + ntoy(10);
+        emulator.update_context(bob(), 0);
+        emulator.contract.ping();
+        assert_eq_in_near!(
+            emulator.contract.get_account_staked_balance(bob()).0,
+            deposit_amount + ntoy(10)
+        );
+        emulator.contract.unstake_all();
+        emulator.simulate_stake_call();
+        assert_eq_in_near!(emulator.contract.get_account_staked_balance(bob()).0, 0);
+        assert_eq_in_near!(
+            emulator.contract.get_account_unstaked_balance(bob()).0,
+            deposit_amount + ntoy(10)
+        );
     }
 
     /// Test that two can delegate and then undelegate their funds and rewards at different time.
